[roll] Roll fuchsia [system-update-committer] don't commit if current is already healthy
In this change, we modify the system-update-committer so that if the
current slot is healthy, we don't need to commit. Having it commit when
current is healthy (the previous behavior) isn't functionally any different because the paver calls are just no-ops. However, it makes the code complicated.

To optimize for simplicity, let's change it so that when current is
healthy, we need not commit. Also, we can rename things to make them
more intuitive.

Test: fx test -o system-update-committer-tests
Test: fx test -o system-update-committer-integration-tests
Test: fx test -o system-update-checker-integration-tests
Test: fx test -o omaha-client-integration-tests
Original-Fixed: 67237
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/468134
Original-Revision: 1bc6ac11f394e2be5198af2890e74e78286ddde9
GitOrigin-RevId: 59223f2987980018ac64c47e9f0e94a743725576
Change-Id: I5c7d74a45b1b3e1ed7e841e7845367e96595a292