[roll] Roll fuchsia [bootstrap] [fshost] Add decompression sandbox component for blobfs.
This will not enable the use of the decompression sandbox component, but
add it for use by the fshost component. Until it is flagged on by blobfs
the component will not get started up since no connection will actually
be made to it.

Original-Bug: 62177
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/463820
Original-Revision: cbdaffe7bfa2a15eaea01bd67bb232ffb7c98be6
GitOrigin-RevId: d85b57c1dffa8485a066d5827e9efbfed7fe1d0f
Change-Id: I1beed5241cf8bfe6c4835a191b6ea8da3fb85977