[roll] Roll fuchsia [kernel][debug] add command to display build instrumentation.
This CL adds a k build_instrumentation command, which prints a list of
the build instrumentation features available in this kernel build.

The list is non-exhaustive as each new instrumentation has to be
manually added.

This command is intended to be used by host tests so they can check if a
given instrumentation is enabled or not before running a test.

The output looks like this:

```
$ k build_instrumentation
[00024.594] 01652:02533> build_instrumentation: address_sanitizer
[00024.594] 01652:02533> build_instrumentation: debug_assert
[00024.594] 01652:02533> build_instrumentation: done
```

BUG=66578

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/464984
Original-Revision: eed5375fbbdc8bce1560125e68a56cd2075642e2
GitOrigin-RevId: 8344e740ca1df5d6e1eec108f6f2652922f97deb
Change-Id: I99ea4e41cf8854ea3894dd4858a86bca24fa9dee