[roll] Roll fuchsia [component_manager] Add a Discover action
This adds a Discover action, and dispatches the Discovered event under
this action. This allows us to guarantee that Discovered is always
dispatched before Resolved, and is never double-dispatched.

As a drive by, remove some errors around invalid environments that
aren't actually possible and replace them with unwraps. The errors in
question arise if something is wrong in the environment FIDL, but this
shouldn't be possible the FIDL is already validated (which we assume
elsewhere).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/463604
Original-Revision: 8eafb4d6cb79bfeb3b037612143a854d7f55e360
GitOrigin-RevId: a5bcc70e302f8d02ab1d32b38e90a3ed344a1c4c
Change-Id: I208369690266078c9e220b1edb28c4c5958e8d48