[roll] Roll fuchsia [omaha-client] unify representation of consecutive_failed_update_checks
Previously OMCL contained one value of this stat in storage and one in
the Policy, which were not kept in sync. This change loads the value
from storage into Policy at startup and keeps the value in storage in
sync with the Policy value throughout the execution of the state
machine.

Original-Fixed: 66707
Test: Added test proving that the value stays in sync between storage and memory
Test: Added new field to existing Context persistence tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/468795
Original-Revision: 135c122764c96cb5132d4d1bbb51bb75b2f12a92
GitOrigin-RevId: cb6d40c11fee4b81c0360ccd2abae7df01e091fa
Change-Id: I3a1e96e73b02401d88073890b592820e041d0a19