[roll] Roll fuchsia [kernel][libc] Fix printf bug
We were treating precision as width - and if both width and precision
were specified, the latter would win and the former would not be
applied. A simple manifestation of this bug was the inability to apply
width to string_view printing.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465616
Original-Revision: a053b5f494c5866cc31dcb82d753ae33654324e8
GitOrigin-RevId: f53c1155dc19eed0d9adf2e2385137dec76eac1f
Change-Id: I87f7e8347f407e9947f52def0f14fff14a3b1a80