[roll] Roll fuchsia [fshost] Move test into fshost.
zircon/system/utest/fs-recovery contained an fshost integration test.
This CL moves it into fshost proper, and makes it work in a v2
component.

To make this work, I had to refactor the options a bit so that they're
in their own class and add support for the boot-args options.

This should also fix the flake mentioned in the referenced bug.

Original-Fixed: 66948
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/466654
Original-Revision: b3b92ce9ebdaa284af8262675df6a0ed28b48842
GitOrigin-RevId: fce45f2dfcced4ced0ea71040d104dbea1422e36
Change-Id: I13f9ea49c2a1974e1a7bfe9e23187a50f93a58a2