[roll] Roll fuchsia [gidl][c] Test FIDL C API with GIDL
This CL introduces conformance tests against the C
fidl_linearize_and_encode API generated by GIDL.

The motivation for creating these is that the LLCPP API
(already being tested) may conditionally change which
C API is being used based on performance considerations.
This already happens in cases where the message is an
inline object that does not need encoding, but in the
future it will also conditionally switch between
fidl_linearize_and_encode and fidl_encode_iovec.

The idea is to add more C APIs under this GIDL target
as needed to ensure coverage and correctness, regardless
of what LLCPP is doing.

This CL is reusing the LLCPP object builder, which builds
objects compatible with the C API, but it may be desirable
in the future to create a dedicated builder for C objects.
fidl_encode is not tested in this CL specifically because
it requires the object to be linearized before the call
takes place.

These tests could potentially replace some of the tests
in encoding_tests.cc and decoding_tests.cc at some point.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/468534
Original-Revision: b85d04a40f29b689c25c45a944236edcfb65703b
GitOrigin-RevId: 5bd8c1a42a17e0252ef68d6df2734b4cdc56517d
Change-Id: I0e4c300c44ab6858a75fb1e2a7ef548730cad71c