[roll] Roll fuchsia [runtests] fix invalid memory accesses.
The arg parsing logic in runtests had an off-by-one error. Valid
arguments from argv are in the range [0, argc).

When the last argument was a flag with an extra parameter (for
example `-n`), `optind` would be equal to `argc`, however the check
only compares whether `optind` is greater than `argc`.

The resulting logic was that if a flag was missing, most of the checks
would pass regardless, and garbage data would be read.

To reproduce this bug, you can run `runtests -n`, without an extra
parameter.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/469197
Original-Revision: d2e077a889ef6acc01a12b5687eb8579a86ee03f
GitOrigin-RevId: db8a042578b9502d7599926b482ab62405e2b152
Change-Id: I5c10002869e57cc4be46d1d684ecf1a543ebeff9