[roll] Roll fuchsia [fshost] Add build dependency for enabling the decompression sandbox
Adds in the option for sandboxing decompression to the fshost config and
plumbs it through to the mount options.

This currently adds a small rule that also includes gpt_all, mainly to
support x64 which currently uses this option as well. This will probably
change anyways when I actually go to enable this for something.

Original-Bug: 62177
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/463821
Original-Revision: beb5e13bd04ed5e21d855ec69ecfc94e2dca12a4
GitOrigin-RevId: 4a081721de18270bc6270f27a0e94894c097ebe9
Change-Id: I68195e3489b25ba9b1621468f5c6eac98b4553c0