[roll] Roll fuchsia [proc][bionic] Add toolchain and declaration
This CL addins a GN target for building bionic in the Fuchsia source
tree. We do not currently have the source code for bionic in the Fuchsia
tree, but we are working towards adding it.

This eventual goal is to be able to run the bionic test suite against
Fuchisa to validate our *nix compatibility.

Original-Bug: 66359
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/468632
Original-Revision: 23ff09ad0b0e3e3666b0ef8d24cc6a320382b0e5
GitOrigin-RevId: 72bd58f3479e7c3aa73aafceb22f94ebac6c5ad9
Change-Id: Iee639d3630bae6e6c1276e734a176efa2f071a03