[roll] Roll fuchsia [far] Implement FAR validation in the C++ reader.
This changelist contains a port of the FAR validation functionality
from the Go and Rust readers to the C++ reader. It includes unit
testing for the dir name validation function and uses the same invalid
FARs to test the validation functionality as the Go and Rust readers.

Original-Fixed: 55804
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/461994
Original-Revision: 37837cb90a5c42522db9e1c287a5934f3346959e
GitOrigin-RevId: aeb25c87d08e228ad8e0f9e9ce035051d20c16e2
Change-Id: Ifb8e77f50ae59def28f9be3404d53b28d8fe8d39