[roll] Roll fuchsia [fxl] Fix invalid use of volatile in C++20
In C++20, volatile variables of non-class type cannot be used with
compound assignment operators (++, +=, |=, etc.) and also cannot be used
as the left-hand-side of an assignment where the result of the
assignment is used in a larger expression.

The latter type of usage appears in an FX_DCHECK in the destructor of
fxl::WeakPtrFactory, which has now been broken out into its own private
member function to satisfy the requirements of the C++20 standard.

TEST=no behavior change

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/461725
Original-Revision: 5b6dfa94de2d4ac2f04ae6272a69de2cdff0c099
GitOrigin-RevId: 19eb862c5725228243e572accf468ed3c6a7811b
Change-Id: Iaf9ae64a9ff7c702f0d902df783640150953fcc3