[roll] Roll fuchsia [fint] Add gn_trace_path to set_artifacts.proto
This needs to be done as a separate commit from the commit that starts
setting the field – otherwise recipes will fail to deserialize the
textproto because the textproto format does not allow unknown fields.
We'll have to wait for this commit to be copied to the recipes repo
before starting to set the field in fint.

Original-Bug: 65899
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/470839
Original-Revision: 3be7e4d93120710075380828211be04591dfd594
GitOrigin-RevId: b87bc7df49bf816d0a918d05f7558b1a7b5132a8
Change-Id: Ie0820e51e9881c53319f9fb845139edb7b02272a