[roll] Roll fuchsia [bt][hci] Fix race condition regression.
The previous CL introduced in a race condition in the test when fixing a memory leak.
The Event Arc must be consumed when it is used as a raw pointer in the
test to avoid it being double freed.

Test: Run bt-hci-lib-tests
Original-Bug: fxbug.dev/67595
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/470821
Original-Revision: 2fdc94108c831cbb25295be489ee5e206d764ac0
GitOrigin-RevId: fe9313ea292c54ae383b470b3fdb70a277387ab2
Change-Id: Id27a12f645348c885534357b216294a600a65d2d