[roll] Roll fuchsia [kernel][vm] Add inactive pager backed queue
Introduce a new page queue to track inactive userpager backed pages.
Pages are moved into this queue when they are explicitly promoted for
eviction, via the promote-no-clones command line for now, and with
eviction hints going forward (not yet implemented).

Pages are moved out of this queue when they are evicted, or when they
are accessed again which moves them to the first LRU page queue.
Inactive pages are considered for eviction before pages in the oldest
LRU queue.

This change only has effect if the kernel.page-scanner.promote-no-clones
command line is enabled. Instead of moving the promoted pages to the
tail of the oldest page queue, these pages will now be moved to the head
of the inactive queue, maintaining an LRU scheme across all inactive
pages.

Test: kernel-unittests

Original-Bug: 65334
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/470430
Original-Revision: 1e32808904337ffd412c3d680a928f7bbbf8749d
GitOrigin-RevId: fbf4405ff062851382720dc88985284acbd9752b
Change-Id: Ic79804045d99437205f4056aa8824541d3db6d3c