[roll] Roll infra/recipes [fint] Add gn_trace_path to set_artifacts.proto
This needs to be done as a separate commit from the commit that starts
setting the field – otherwise recipes will fail to deserialize the
textproto because the textproto format does not allow unknown fields.
We'll have to wait for this commit to be copied to the recipes repo
before starting to set the field in fint.

Original-Bug: 65899
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/470839
GitOrigin-RevId: fba450066186e56ab4b17053d2d3185a749ecb0e
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/470859
Original-Revision: 8a084b0269615fbc56c3bc596dc0f77ecb56bc0b
Change-Id: I2377f5263b27a03b8385abe6d84b6fbfa155c8ef