[roll] Roll fuchsia [flatland] Remove Validate Function
The Validate() function is no longer necessary
for the Renderer API and so it is now removed as
entirely, both as an API level function and as an
implementation detail for the subclassed renderers.
As a result, the BufferCollectionMetadata struct is
no longer needed as well and is likewise removed.

Testing: Existing tests are updated to account for
the API change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/466155
Original-Revision: d5e2fc8b8bc3d683c3ef003a5873e07fc0077826
GitOrigin-RevId: 9589e6584b890ba05c32d79cb01ca9846421a575
Change-Id: I5cc09ed91d99bdc2738c9efe9091f4e52b8cd4b0