[roll] Roll fuchsia [fint] Pass GN trace path via set artifacts
Instead of writing it to a known path that needs to be hardcoded in both
recipes and fint.

For now, we'll keep writing it to the same path in the build directory
to enable a soft transition with recipes.

Original-Bug: 65899
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/469601
Original-Revision: 8b64ef8e103d2b1c66095219d41540b44cf131d3
GitOrigin-RevId: 4bc07c2425e4de9803e88328fe982f4fef55a129
Change-Id: I374563a79681daa3f79358db800599cce31ffe0a