[roll] Roll fuchsia [httpsdate] Calculate poll offset from system UTC clock
Calculates, but does not yet report, the offset of the time polled from
the server and the system's UTC clock. This will serve as an
approximation of the error distribution for a single poll. Eventually
these metrics will be used to improve HTTPSDate's estimate of sample
error.

Original-Bug: 63370

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/468678
Original-Revision: 5903391dbf06dc507af3465fabafbd9a1017956e
GitOrigin-RevId: 3ab5999757cbfe7893bf71ca81092e59372990ee
Change-Id: If1e5a757ddf4538db565f10c763650f40bf444c8