[roll] Roll fuchsia [settings] Allow policy handlers to intercept and modify setting events
Policy handlers need to be able to intercept Changed events sent from
the setting handlers to the switchboard in order to adjust their values
according to policies.

When the policy logic was first designed, the Changed event caused the
recipient to send another Get request for the value. Now, the Changed
event contains the updated value, which eliminates a round trip, but
also was not intercepted by the policy layer.

Original-Bug: 67148
Test: fx test setui_service_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/469214
Original-Revision: e3b2795620c0b0deaff601321922dee9adbe6462
GitOrigin-RevId: a96edf8d386bfe0979d0a08517aaab1efdae0999
Change-Id: Ia1d8c94cee794c78688211674077f20c7d9ec3b5