[roll] Roll fuchsia [bindc] Add support for the new bytecode format
This CL adds an "use-new-bytecode" arg to the bindc program.
When the arg is true, the bytecode produced will be in the new
bytecode format.

The bind compiler is refactored so that the instructions
and the symbol table are wrapped in a BindProgram struct. In
addition, the logic for encoding the instructions into the
old bytecode format are moved into a separate file. A skeleton
file for the new bytecode format logic is added.

Tests: Added a unit test for argument
Original-Bug: 67440
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/469675
Original-Revision: 5fa3e644fd78457e2f72414539b15a66e6e20a37
GitOrigin-RevId: a76dff8c38d25e1b68bb2956968794c351f40f4f
Change-Id: I97ac01549061e24a0d426afa79277af566fa441e