[roll] Roll fuchsia [debugger] Split out and enhance exception docs.
This splits out the existing two large documentation files into
topic-specific files. This will make it easier to modify and read.

The documentation hasn't really changed. Some parts of the documentation
should be enhanced and this is more obvious from this move. we can
enhance it in a separate pass.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/467955
Original-Revision: a5ac729382158559717930e18e6d610d38829977
GitOrigin-RevId: 5c90cda708485b08601316243ceaa3c255865151
Change-Id: Id115e57ff2242e0b213406bccca636b0b4524db1