[roll] Roll fuchsia [fostr] Support for printing uint8 and int8
This patch introduces support for printing uint8 and int8 members (for
structs, unions, tables, and vectors) in an ostream<< context.
Otherwise, these members are interpreted as a C-lang `char` type. Note
that `char` is not a FIDL primitive.

This patch also adds two small tests to verify arrays of primitives are
formatted correctly, such as array<float, 2> and array<int32, 2>. The
user can read the tests and thus look for the source of compilation
errors elsewhere (such as a missing `using` statement).

BUG: 67157

TEST=CQ

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/469134
Original-Revision: 0d88bd6bc31ca93571aa647e206785dc0a50844d
GitOrigin-RevId: 51fd5f5980cea66273a1d0a7a621c78500d57cce
Change-Id: I72ca96e4cb8c56a8fa8aa42b7c7f16ed011067da