[roll] Roll fuchsia [bt][a2dp] handle session publishing failures
In some situations, a Session2 publisher might fail or not be available.
Continue trying to play the audio without a session id (using zero) and
skip starting the AVRCP relay to the session info when that is the case.

Original-Fixed: 67502
Test: added sink_task_builds_without_session
      fx test bt-a2dp-encoding-tests

Multiply: bt-a2dp-encoding-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/470077
Original-Revision: ae95db914231cfc6314950ba2d97b9319f421098
GitOrigin-RevId: 5603a4172ecb23871c2038ab7339c968457ed532
Change-Id: I12bab338b16dc106094940edf88fbe50c9e594eb