[roll] Roll fuchsia [brcmfmac] Improve and count log throttling
With log throttling in place this change adds the ability to count the
number of times a log entry was throttled before it was logged again.
This is done by appending a string like '[Throttled 3 times]' if the
message was previously throttled, if no throttling occurred nothing is
appended.

This change also replaces the old throttling functionality that was
based on a counter to use the new throttling method.

BUG: 29355, 63984
Test: Included unit tests, manual testing with logs enabled/added

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/470521
Original-Revision: feedc246b03e75029a064d34068cb20ab2bd98fa
GitOrigin-RevId: 2e47c6163e79d51897e4ebcfa96219a0d6e090c8
Change-Id: Id7704bfbe01df7b56e84b76de47e76a602d24b53