[roll] Roll fuchsia [fint] Use JSON instead of textproto for set_artifacts
We use JSON instead of textproto for this message because it passes data
from fint back to the caller, but the source of truth for the proto
definition is in fint. Proto libraries don't have good support for
deserializing unknown fields in textprotos (unlike JSON protos), so if
we used textproto then we'd have to make two separate fint changes every
time we want to start setting a new field: one to add the field, then
another to set the field, which can only be landed after the updated
proto definition has been propagated to all consumers.

Serializing to binary rather than JSON would also work, but the
disadvantage of using binary is that we'd no longer be able to
arbitrarily delete and renumber fields of set_artifacts.proto, whereas
we *are* able to do that with static.proto and context.proto. This
discrepancy could get confusing – I expect it would be easy to forget
that it's not safe to arbitrarily modify set_artiacts.proto fields.

This is not safe to land until recipe change fxrev.dev/470803 lands.

Original-Bug: 65899
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/470843
Original-Revision: 0e3234bd6d2cea89a3e32aed411af5ca6e30540b
GitOrigin-RevId: c243edc2a59bfd9bc4642e26715775a7e273cfb1
Change-Id: I070522e4f03583461bf656233cd6f3f13f3ce140