[roll] Roll fuchsia [build] Ensure all shared_library() entries are listed in binaries.json
This CL fixes an issue where binaries.json was not properly listing
specific shared libraries, like libzircon.so.

The root cause of the issue was that the corresponding metadata
'binaries' key was not set on the shared_library() target that actually
builds the library, typically in a -shared companion toolchain (e.g.
'//dir:foo(//build/toolchain/fuchsia:x64-shared)') but instead on
the redirection target in the base toolchain (e.g.
'//dir/foo(//build/toochain/fuchsia:x64)') which is a simple group().

Build unification changed the way some sysroot and kernel libraries are
referenced, as a consequence, the redirection group() targets are no
longer in the build graph for these (instead, the -shared targets are
addressed directly), and the metadata was thus missing.

This fixes the issue by moving the metadata key definition to the final
target, which also avoids making a lot of computations related to the
target's final outputs in the base toolchain.

For testing, the content of binaries.json after 'fx gen' is run
(no need to build) was compared. This shows new entries for libzircon.so
as well as libc.so when a variant is being used (which isn't 'asan'
since its variant is always built for the SDK anyway). Note that
"label" values have an adjusted toolchain path too.

Original-Fixed: 67627
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/470877
Original-Revision: 33c789deff7a098ea229b2e9d5cae595561a0602
GitOrigin-RevId: 828574f853aceac7d500b464b25e309fb7feba77
Change-Id: I48710bc06a3b9096fead3d60c9e2bf962b92967d