[roll] Roll infra/recipes [testing] Rename stdio log if empty
Due to crbug.com/1130788, Logdog shows a confusing "no such entity"
error message when trying to view an empty log. To make it more clear
that this is because the log is empty, rather than the result of a
transient server-side error, add an "(empty)" suffix to link names for
test stdio logs that are empty.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/471978
Original-Revision: ca74800f0ff94f10619d0d3ef78de917316c3ee7
GitOrigin-RevId: 18866d9212f2a8ed61c01c21f3047b1b0a72fb8b
Change-Id: Iefa5d260ea670f352dc4642a0606bbe9c9a05ec8