[roll] Roll fuchsia [ffx][mdns] drop tasks for lost ip addresses
Linux does not error on sendmsg / recvmsg when the socket is bound to
an address no longer on the system, and worse even allow packets to
still be sent with the bindings source address (effectively spoofing
the source address). In order to avoid this outcome, explicily remove
tasks for sockets that are bound to no longer present addresses.

Original-Bug: 67158
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/471719
Original-Revision: c5a77391796aa22c4bc9fabed3004be7f7362da6
GitOrigin-RevId: 6e00e7ed2a2eba244f08452b858b3d98f2704332
Change-Id: Ibf4204ef556badade2db8251f2f0e9a246c07561