[roll] Roll fuchsia [location] Send initial null region code
This change is the first of a couple changes to store and send the
initial regulatory region code on start up. If no region code is loaded
on boot, a null update would be sent.

This change adds the ability to send a null update through the API. Only
the first update sent can be null.

Test: fx test regulatory-region-tests
      fx test regulatory-region-integration-test
Original-Bug: 66538
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/469604
Original-Revision: 9db6edc114cf1ea1128a4bc54ead2d4b16347cae
GitOrigin-RevId: cb5fc7ccb77997f66de7b4cabf7872225c2d945e
Change-Id: Idc2b2acdba6d0e709fefeaf5529b6357f42e7421