[roll] Roll fuchsia [swd] send errors to the syslog
Returning an error with `main() -> Result<(), Error>` emits logs to
STDOUT. Unfortunately STDOUT isn't sent to the syslog, so when these
processes are launched as a component the error is lost. This changes
`omaha-client`, `pkg-cache`, `pkg-resolver`, `system-update-checker` and
`system-update-committer` to also emit errors to the syslog.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/468902
Original-Revision: 38552c2cd973ee422a1962cc5c1ef4af698aa364
GitOrigin-RevId: f693138dcbdac13738b9504dd4190378c770d051
Change-Id: I384f747deb18330faa02227236c6567e9c848a85