[roll] Roll infra/recipes [affected_tests] look for paths of host tests
The ninja dry run output doesn't reliably contain stamp files of the
form that the script is looking for. The exact pattern of what to look
for depends on the specific GN templates used to define the test, but
all the examples I've looked at include the path to the test executable,
so search for this.

Original-Bug: 66869
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/471557
Original-Revision: fcbc1df2f402a7e9888d59a31c232ccd9f952162
GitOrigin-RevId: 9321b1d84a814492afc56550ac467fa3b343c3cd
Change-Id: I60374203130f1dfdb153c468aab3934ee4ab9f83