[roll] Roll fuchsia [BUILD] Template for prebuilts with flavors (variants)
This is a template for selecting which flavor to use for packages that
are prebuilt into multiple flavorss that need to have the same runtime
component URLs.

It allows for the definition of multiple flavors for a prebuilt
package, and a build param which is used to easily allow a developer
to switch between those flavors.

As this can obscure which flavor is being used (the flavorss are no
longer distinguished in the build graph or component urls), the template
generates a manifest of what flavor has been chosen for each package,
so that build outputs are auditable.

This is meant to be transitional, while other, more comprehensive
tooling, is created to facilitate these sorts of needs.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/459507
Original-Revision: 65e628f2b28fb36278166b248cd6738149916d3f
GitOrigin-RevId: 30aae3e1726bc251ada95b3394a32cb434ce7be4
Change-Id: Ie773bc2dabd9aa813c71627c9900e2b075106157