[roll] Roll fuchsia [tee][ddk] Replace tee.Device with tee.Application
The TEE banjo protocol essentially just serves as a connector to the
fuchsia.tee.Device FIDL protocol. The fuchsia.tee.Device protocol is
being replaced with a protocol that is scoped to a single trusted
application instead of the entire TEE. In order to appropriately scope
the channel, the client must now provide the trusted application's uuid
when the connection is first established. This CL adds a new method to
the TEE ddk protocol to provide the UUID and establish a
fuchsia.tee.Application channel. It also updates the clients to use the
new protocol.

Original-Bug: 44664
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/403569
Original-Revision: d27d6d3ccad4d75e6c5817f9f415758c0881bb67
GitOrigin-RevId: e053314904cb753fccd7eebd102ece19a22dd066
Change-Id: Ia9a7abc2977fff50f1182e628ecb06b80fda8eec