[roll] Roll fuchsia [stdcompat] Port fit::string_view.
This implementation has a couple of buggy methods, that don't match the
specification. The issue is that the tests are incorrect. To simplify
review, this CL is merely getting this to compile in stdcompat.

The tests were moved to gtest, which has two effects:
 (*) Literal integer type must be specified. zxtest allows integer types
     to be promoted to a bigger one for comparisons.
 (*) EXPECT_STR_EQ -> EXPECT_STREQ.
 (*) EXPECT/ASSERT(args, message_format, message args) ->
     EXPECT_ASSERT(args) << message ags;

Original-Bug: 57349
Test: stdcompat-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/467595
Original-Revision: 7787c93a4a91882447b18b75c5b645fd40b94fd6
GitOrigin-RevId: b87134567a10b515539575b0494a6f4c88cc70e2
Change-Id: I539fe77a70a8562705921a705fff075fe6c26294