[roll] Roll fuchsia [session] remove non-working input_session test and re-enable tests
As discussed in fxrev.dev/472069, the input_session test wasn't
compiling, and build targets had been ignoring it. The test, as
structured, was not relevant to the example binary, so it has been
removed, and replaced with the default "TBD" stand-in test.

Build targets are updated to re-enable the stand-in tests for
input_session and scene_manager_session examples.

Also renames the widely used noop test placeholder, and updates
references to `fx run-test` (deprecated in favor of `fx test`).

Original-Fixed: 67160
Original-Fixed: 67161
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/472159
Original-Revision: b9ead271571af685d549f36016c28908b695baac
GitOrigin-RevId: 4cb341bc0a236b6382e6333a2c6df5417636a77c
Change-Id: Ie37b036355243fe9d3cc84bf6ae36a29329487d6