[roll] Roll fuchsia [location] cache and load region code
This change adds a cache for regulatory region services to store
the region code when it is configured. When the regulatory services
start up, the initial value is loaded from the cache.

The cache may be deleted if disk space is almost out. Regulatory region
services does not rely on this data to necessarily be saved; it still
expects to be configured at startup if needed.

Original-Bug: 66537
Test: fx test regulatory-region-testst
      fx test regulatory-region-integration-test
      manually tested that region code is loaded at startup after
      rebooting.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/471023
Original-Revision: 649b3884d117a9a822851391680b8a728fb22879
GitOrigin-RevId: 766003f460d3e62e056d8fa32455215e4fbefa38
Change-Id: Ib144baef182bdc67351201ee7a013fdeefa9768b