[roll] Roll fuchsia [fdf] Clean up ETH_PHY platform defintion
This value is only used for Realtek drivers and it was put in the
fucshia.realtek.platform bind library, so we should make it clear that
it's for Realtek in platform-defs.h too. Also change the value to 0x03
to line up with the other Realtek DIDs.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/472521
Original-Revision: c4d2de18aa4de8a028b768dd6786715bfb96abcb
GitOrigin-RevId: 8bd34e4fe4178e8db896a9127d640b2cd2e64c10
Change-Id: I5e6ed6f1fef2b7bfb8d773a4c4d083690574c708