[roll] Roll fuchsia [driver_manager] Add support for loading fallback drivers sooner.
Currently when booting from a USB, the system hangs during early boot
because the `usb_composite` driver is a fallback driver, which means it
won't get bound until system drivers have been loaded - but system
drivers are on a USB device, which requires the `usb_composite` driver
to be bound.

This CL adds a flag, `devmgr.bind-eager`, which is a comma-separated
list of fallback drivers to treat as non-fallback.

Test: fx test ddk-fallback-test
Test: Boot from USB on NUC no longer hangs before minfs/blobfs comes up.
Original-Bug: 44586
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/471837
Original-Revision: 4eba6d7ba8efee2ebf3927d87fc7b8053c40fb1b
GitOrigin-RevId: 357cf9cd5e1bd6783fd6af7927b197e9b229cddf
Change-Id: Ia30eb130f86bdfb9f4f1a75e937dc8351d0ec111