[roll] Roll fuchsia [docs] Freshen up docs/concepts/framework
Rename framework to process, to create a distinction of usermode
framework things that are ancestral to components.

Remove boot_sequence.md, it is hopefully outdated and the topic is
better covered anyway in
everything_between_power_on_and_your_component.md. Plus users have been
complaining that the stale contents are confusing to them.

Deleted hub as it's a components concept, the page contents were out of
date and not super helpful anyway, and the details vary between v1 and
v2 components so it's irrational to introduce this concept at this
point.

Deleted some links from markdown that's not under docs/. We don't care a
whole lot about markdown that's not on fuchsia.dev these days. Relics of
bygone days.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/472914
Original-Revision: fb05da412bc57c9f02ff4625252c1f10fab99295
GitOrigin-RevId: 4e3f2e7b79210d61c742e0c261d4cf792b0756d9
Change-Id: I9a75bf10fdcd9fe17720c0115a7cacf685a1d5be