[roll] Roll fuchsia [bt][a2dp] Initiate streaming if the peer doesn't
If the peer connects to us and doesn't start the stream, try to start
the stream using our normal method.

The delay is the same as the initatator delay.
Skip the delay when we establish the control channel.

Original-Fixed: 50465
Test: added tests; fx test bt-a2dp-lib-tests
Test: connect to device manually and stream w/phone
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/471462
Original-Revision: 5d4c121505f87d416ee5d20296790a0e2b4bf8b1
GitOrigin-RevId: e3d4b3e53c5f800cdf7c32035b1f7783bf9093f8
Change-Id: I1337f159edf0cf3b3e331600d5bad89894e03efc