[roll] Roll fuchsia [docs] update kernel debugging documentation
I updated the instructions to connect to use gdb to connect to qemu.
The old instructions didn't mention that the usage of the gdb script
was mandatory (for kaslr relocations). The new documentation
explictly mentions that you should use the script for using remote gdb
then it mentions that it is possible for it to load automatically (if
set up correctly), and notes tha it won't work right away with qemu-kvm

BUG=67893

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/472417
Original-Revision: adf66556c91da857fa699c8bc0693813edb80693
GitOrigin-RevId: 64c1ab335004d5ebee47917ab180d5cee54ba2fc
Change-Id: I047f43e0438af41536f2749b3ae0234c6c84c038