[roll] Roll infra/recipes [libxml2] Do not use CMAKE_INSTALL_PREFIX
Building libxml2 currently fails during the configuration step due to
CMAKE_INSTALL_PREFIX being an empty string that's passed to
`file(RELATIVE_PATH ...)` which requires an actual file path. This
updates the cmake invocation such that CMAKE_INSTALL_PREFIX is set as
the install directory and we do not set `DESTDIR` when installing.

Ideally, this would be addressed upstream in the cmake file. This is a
workaround that provides the same behavior that we currently have.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/472851
Original-Revision: 4b94c178e5ffee145b68542a70540d006808f2b6
GitOrigin-RevId: 1227fc64d79ace51826b624c1e6ca812aecae623
Change-Id: I570d42ffa6f7ed73db64085f7c1e03872630541b