[roll] Roll fuchsia [vfs] Add initial pager capability to C++ VFS.
This adds the initial skeleton of pager capabilities to the C++ VFS. The
long-term goal is to replace blobfs' pager framework with this one, and
the even longer-term goal is to remove the non-pager paths from the VFS.

This patch implements the simplest possible read support and a test that
runs against it. There is a lot of work left, especially around closing
files and shutdown. There are also a lot of needed test cases left, some
of which are listed at the bottom of the pager_test.cc file.

Original-Bug: 51111

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/466474
Original-Revision: e152ffb82060d15f07384cfc280ab26d7596216d
GitOrigin-RevId: b9b5ae102937872662546760ad1a7dc907329263
Change-Id: Ie64fd106e852bfa650736b44e2dd8c4122296274