[roll] Roll infra/recipes [build] Wrap entire build in macos_sdk()
Any step that runs GN or ninja and might run on a Mac builder must be
wrapped in `with api.macos_sdk()` to ensure that XCode is set up, since
the Mac build assumes XCode is ambiently installed. This was not the
case for the should_build.py script, which was causing failures on Mac
builds.

For safety, also wrap the entire `nest("build")` in macos_sdk() to
guarantee that all sub-steps will always have XCode set up if necessary.
This also lets us reduce the number of times we have to set up XCode by
storing a boolean in the macos_sdk module to track whether XCode is
currently enabled, and skipping the setup steps if it's already enabled
at a higher level in the call stack.

Original-Bug: 65899
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/473301
Original-Revision: 86491c94a14f7752b39c0e33e66fff7f9e3a12ea
GitOrigin-RevId: 2785285645752c836056003ae30791f03063d371
Change-Id: I6f987635463cdd0ac15501011f6c8eb920aa3cc5