[roll] Roll fuchsia [camera][lib] image_io_util allows caller to specify dir_path
This CL changes the util's behavior to allow callers to specify the path
to write frames to, replacing the previous hardcoded solution.

Example usage of the util can be found in the unittest, which writes to
the on-device cache.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/471637
Original-Revision: 72aaee82b5b8bccafc53ec022a1a76a7c2806c03
GitOrigin-RevId: 43727defd67bf29db9d97ebdc5bc141ef54a6e76
Change-Id: I21acb33fd21c1bb066c324a45fc5700b35b81c7f