[roll] Roll fuchsia [build] Remove code paths related to 'zx' variable.
Since the 'zx' variable now always expands to '//zircon',
we can remove all code paths that were specific to the
Zircon build, in other words:

  zx == '/' is always false.
  zx != '/' is always true.

NOTE: This does not remove the definition of the zx variable
yet because it is still used widely as $zx/ in lots of build files, but
this will be addressed by a future CL that will do a mass string
substitution. Also, some vendor/ and third_party/ BUILD.gn still rely on
it.

+ Move zircon/public/gn/test/zbi_test_success.gni to
  //build/testing/zbi_test_success.gni

Original-Bug: 60410

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/473238
Original-Revision: efec64a885d587eaf55a0cc0b02621db57448a36
GitOrigin-RevId: e7b7de06a0b7ce7604a2ef7b3c6f7cc0ec665abb
Change-Id: If47228cbf52a036f967d078d5ff2385533fef42a