[roll] Roll fuchsia [fidl][hlcpp] Split golden tests
This splits the HLCPP golden tests into two tests, one for the
traditional bindings and one for the "natural types" variant. This
allows the goldens for a given test to all be in the same directory
which makes the build rules (and the changes I'm trying to make to them)
much saner.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/471318
Original-Revision: 771f5d0c19356f595682ea0e9613cff41085e817
GitOrigin-RevId: 8e0cbccf9e7f8b2feb0d080bf99e01cd7315c5ce
Change-Id: I39dca418c646c8253901136025bb942ce5efa089