[roll] Roll fuchsia [httpsdate] Only record offset for the second poll
HTTPSDate attempts to schedule each polls such that the bound on UTC
time it has is halved by the poll. This is achieved by scheduling the
poll such that either the minimum or maximum UTC possible with the new
poll falls in the center of the accumulated bound. After a few polls,
the accumulated bound is very small. When HTTPSDate schedules another
poll in this situation, the center of that poll is very close to +/-
.5 second and skews our metric.

This change updates HTTPSDate to record only the second poll to remove
this effect. The first poll is not used as it usually takes longer due
to setting up TLS connections. The second poll is likely still fairly
random as the algorithm has not yet tightened the bound.

Original-Bug: 63370

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/472217
Original-Revision: f6c9289ed7397f68ef9b10f3cd0d10b245af4cdd
GitOrigin-RevId: 9306f8097e810889cf2950c8fcaa3b83bd81ffaa
Change-Id: I3c2d90a6065d30887a2a68e2ca6e45260c7c5404