[roll] Roll fuchsia [fidl_codec] harden handle tests
To test fidl codec, we send a message over a channel and decode the
message received at the other end.

For channels, the driver can change the numerical value of the channel
(and it will change it every time if a restriction of rights is asked).

This patch ignores the numerical value of the received handles and
replaces it with well known values. This way, we can reliably check the
decoded message text.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/473332
Original-Revision: bbee336214fefa7cf31eaaad9fc9dff6a49963c8
GitOrigin-RevId: 0d22783070e87617a37ae2fc4051043fee727630
Change-Id: I4a4b50bc755638ce5d2f9ed00ed01e96642967d9