[roll] Roll fuchsia [fidl][dart] Better exceptions in test bases
Calling a method on a test base class throws an UnimplementedError but
this wasn't helping people actually fix their tests. With this change
that exception includes which method on which protocol isn't
implemented, and tells the developer to implement it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/472992
Original-Revision: f855746c8090cb3fc981bf87e43db5395989ef8e
GitOrigin-RevId: 021688a792c5a3b85dbd652292763347006a49e2
Change-Id: Ib29aa89c13b0f0638deebf5006faf81ee9adb0a4