[roll] Roll fuchsia [kernel][x86][topology] Address cosmetic bug
Today we enumerate dies using the tail of the APIC ID that actually
encodes both the canonical package ID and die ID. Although this ID
results in the same effective enumeration of the dies in the system -
in comparison to enumerating all packages using package ID and in turn
all dies within by die I - it gives a misleading presentation both in
source (in the definition of the APIC ID decoder) as well as at runtime
in the early boot dump of the system's topology.

To address this, we introduce a `Package` abstraction to the topology
enumeration process and update the `ApicDecoder` to identify both
packages and dies. This is a purely cosmetic change to the enumeration
internals and results in the same system_topology::Graph being
constructed.

Original-Bug: 61093
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465897
Original-Revision: 7fcd16397929978143f3f257a66de7818aba2bb3
GitOrigin-RevId: 0a69e01c3edd11aa736e8bc768bc0e7f5aae94d6
Change-Id: If77b0c512d6686b724b588dadf799cc492024e6e