[roll] Roll fuchsia [ping] Fix IPv6 scope id override
Only overwrite the IPv6 scope id if a value has been provided with the
`-I` argument and it is positive. This was preventing specifying the
scope id after a % in the URL from working, as `getaddrinfo()` would
populate the scope id field correctly but it would be overwritten with
a value of 0 unless `-I` was provided.

Original-Fixed: 64465
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/452660
Original-Revision: b595fff5fef64cfc3a278fe463d44d23fb44fd61
GitOrigin-RevId: acfe1d824cc178cd5ba79dc1a5d2a653df17685f
Change-Id: Iba6de39b4bda7a97e5e420315e221cf7df1cbb85