[roll] Roll fuchsia [kernel][sancov] fix kernel sancov variant.
It seems like sometime during the build unification, the sancov variant
bit rotted. The configs were renamed and the sancov rule was not
updated, as it's not being covered by any buildbot.

This CL fixes that, setting the config to the correct value for the
sancov and asan configs.

Test:

```
 $ fx set core.x64 --variant "kasan-sancov" && fx build
```

BUG=68034

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/473286
Original-Revision: 36e5d1dbbd89c054c0dfc5efe4eb2d5776bbea62
GitOrigin-RevId: a27f9a92ef4eb41501941b9e34e7b21d56564ef0
Change-Id: Iaf22565aab509ba1bcbfe9b0e2051ecbd34fe87f