[roll] Roll fuchsia [reachability] Fix flake due to inspect race
Due to the non-atomicity of inspect entry creation, when reachability
monitor adds an inspect entry for reachability state into the
`BoundedListNode`, the sequence of events is as follows:

1. an entry is created with a sequence number only,
2. the timestamp is written into the entry, and
3. the reachability state as a String value with the key "state" is
   written into the entry.

This means that it's possible to observe an entry that does not have
the reachability state string value in it.

The integration test was flaking because it treated tho lack of the
"state" string value as an error condition. This has been fixed by
assuming the IPv4 and IPv6 reachability state to be "None" if the
state value cannot be found in the inspect data.

Original-Fixed: 67541
Test: fx test netstack_reachability_integration_test \
              reachability-core-lib-test
Multiply: netstack_reachability_integration_test:30
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/473198
Original-Revision: 5825fd9347bf3d45d9634e22800221106315ed3d
GitOrigin-RevId: 8dbf20b28ba507a291785421b9ffb461e49095a7
Change-Id: Iac46e3581d88bc986b94afe749a1f505755bce88