[roll] Roll fuchsia [component_manager] Rename Realm -> ComponentInstance
`ComponentInstance` is a better name for this struct because at its core
it represents one ComponentInstance. It _is_ a realm too in the sense that
it contains links to children, but that notion is somewhat secondary.

Also rename related modules, identifiers and variables.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/472746
Original-Revision: b4669ff987e3fd6121736abd34adbd75315a2c0e
GitOrigin-RevId: ceaa64f214a7d12da945639a36607220a933f031
Change-Id: Id4fa79a254d9bbda6b7be7e528170a145eea663e