[roll] Roll fuchsia [settings] Implement transforms for intercepted changed events
When the audio setting's value changes in response to a Set
call, it sends a Changed event to the switchboard to tell
it to notify clients listening over a hanging get of the
new value.

This CL captures and modifies the Changed event in the
audio policy handler so that Set requests result in
listeners receiving the properly transformed volume levels
based on the active policies.

Original-Fixed: 67148
Test: fx test setui_service_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/471142
Original-Revision: 6b746642003ed3e0e8e4377ba64fd344e04fa27e
GitOrigin-RevId: f712fdd5982376def431513bed5dd4628c6e87b8
Change-Id: I205f8c1529f2b6ad079883c634ed987569e2a63a