[roll] Roll fuchsia [weave] Implement conditional TSM init
ThreadStackManager will now check the configuration value
("enable-thread") before initializing. If Thread support is not enabled
in configuration, TSM now exits without failing when a Thread device is
not found, but any further calls to TSM will not succeed (either
returning an error indicating the functionality is not supported or, for
those calls that don't return a WEAVE_ERROR, returning false or other
appropriate value and provide a guarantee that there will be no side
effects to the call).

Original-Bug: 67672
Test: fx test weavestack-adaptation-unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/471617
Original-Revision: 9e9172169400d2d484584e693f7d7b31ed101c35
GitOrigin-RevId: 137bc0a75fa05cc210649606cbe4d2f3a36b488c
Change-Id: Ief56df7664af03464b56058e05b95c6e4b3e5a1a