[roll] Roll fuchsia [flake][time] Disable timekeeper integration test on emulators
After increasing timeouts on the RTC timeout multiple times, the
timeout is still triggered when running the integration test on
emulators. From the latest run the RTC response is not generated for
a full 200 ms after the timeout.
This flake has never occurred on real hardware, and the test will
continue to be run there.

Original-Fixed: 68048

Testing: dry run, verify test is run on NUC but not emulator.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/473850
Original-Revision: f5030231b1ff114863c100ac3ccd1bb86533834d
GitOrigin-RevId: 0be9899623bb1cfdc510a38b38cd2624874d0fd8
Change-Id: Ia905f5c415934241c8d14cf47d0912e988288114