[roll] Roll fuchsia [brcmfmac] Fix spurious failures in token bucket
The token bucket could sometimes allow additional tokens to be consumed
if the atomic compare exchange operation spuriously failed. This would
then load the old tick and not perform the necessary checks to ensure
the bucket did not go over capacity.

This rarely triggered a test failure, added another test to run for an
extended period of time.

This fix ensures that the capacity is properly checked on each attempt
to consume a token.

BUG: fxbug.dev/68044
Test: Included unit test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474197
Original-Revision: 753b38f2906b24c8beac25f071b2ad5134bef3a8
GitOrigin-RevId: 7a24b3a25a03c544e1f7061728a20700c1c48d10
Change-Id: Icf5c9ed220cbbfb5b778053b8f13f0c7a6d83b4c