[roll] Roll fuchsia [analytics] Executor with thread and timeout
This change add a new implementation of fit::executor, which manages a
debug_ipc::MessageLoopPoll in a separate thread. It also sets a quitting
timeout, such that when the main thread quits, the message loop will try
to quit itself after the timeout if tasks are not finsished earlier.
Note that the timeout might not work if there are blocking tasks
running on the loop.

The analytics_executor_manualtest is an integrated test to make sure
libcurl works well with the threading and timeout logic implemented in
AnalyticsExecutor.

MULTIPLY: host_x64/zxdb_tests: 50
Original-Bug: 65915
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465244
Original-Revision: 2e5c3d22baa3a8f1e4b2bc48f664c75b51fe04a8
GitOrigin-RevId: f5a8c9fee3cf35d31b9dcc242dc4f10a59bfdc35
Change-Id: I1681a0dc9d9dfd8be22972b7bb2182fb12dd97d6