[roll] Roll infra/recipes [fuchsia_cipd_roller] Include rolled tag in summary
This will make each roll's tag show up in the builder page, which is
useful for understanding the history of the roller.

Also make make the "unable to find common tag to roll" text start with a
capital letter and end in a period for consistency with other summary
markdown values.

Original-Bug: 48271
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/474057
Original-Revision: aaa9c18dd6177b48057b05c1c89171dc0c71503b
GitOrigin-RevId: 83eacca45cf1e383e63b456d95db469d7396c0c3
Change-Id: Ifc5b97ba8d50627e2f8de0a31010c82e4245a8f6