[roll] Roll fuchsia [test_manager] One archivist per test
This is working towards a temporary solution to get us feature parity
of diagnostics in v2 tests. In particular this embeds an archivist
inside each test. A future CL will fetch the isolated logs.

Once we have log selectors (blocked by a log buffer per component in the
archivist) this will be refactored so that we share an archivist for
all tests and the need of TopologyBuilder will be removed.

Original-Bug: 67960
Tested: all existing tests are passing + new diagnostics tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/472998
Original-Revision: 4b71af5dc6da7589510cf55698eb6d135789f814
GitOrigin-RevId: df45c134cfe7e990ce6b9b53a2f353caab7e46d7
Change-Id: If29cf216ca26b1ab36cc388568dc96d7c8141b80