[roll] Roll fuchsia [ulib][zxc] Make zxc_unittests runnable on host
There is one test for zx::status that can't run on host because of a
missing member function, but it's very useful to have and much faster
for your edit-compile-debug loop to just go into
out/default/host_x64/exe.unstripped and run the zxc_unittests binary.
It even works with regular gdb so you don't have to attach zxdb to an
emulator.

Of course there is some testing that needs to be done on Fuchsia but the
vast majority of zxc is platform-independent.

TEST=ran the newly generated binary, tests passed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474409
Original-Revision: 7b95aa87115960907831238cce894c213970919d
GitOrigin-RevId: 25a77975dedfed222b49af701870296d4216b4c0
Change-Id: I4b2ecf6c774fa75544f61ca2eff32ddfceee4862