[roll] Roll fuchsia [kernel][vm] Don't panic on DECOMMIT w/ multiple vmar mappings
A VMAR may have multiple  mappings inside its address range.
If a VMAR had multiple non-contiguous mappings and the decommit
vmar op was invoked on mappings that weren't the first one,
the DECOMMIT lambda in kernel would panic, noticing it was
called on a mapping that didn't include the address to decommit.

The bug is because the iterator EnumerateChildrenInternalLocked
walks a vmar's subregion tree and invokes on_mapping on every
mapping in a subregion, not just ones that contain the address
in question. Fix the bug by checking that the address is contained
by a mapping before invoking on_mapping.

Pair-programmed-with: mvanotti@
Original-Bug: 68272 ASSERT FAILED at (../../vm_address_region.cc:638)

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474719
Original-Revision: a9295e73e4455697cca064aa883624dbd149d358
GitOrigin-RevId: 277398d346a5a8926cb1f73f06146e08f95dd6a8
Change-Id: I5ef34bb448f6f2a7bc517dff64151d74b43cbc03