[roll] Roll fuchsia [lib][vmo_store] Make VmoStore return unregistered VMOs
In some cases it may be helpful to operate on VMOs after unregistering
them, or be able to pass them back to a client of the driver. Currently
Unregister() just closes the VMO handle; this change makes it so the VMO
is passed back to the user instead.

Original-Fixed: 68209
Test: fx test vmo-store-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474700
Original-Revision: 541ff78ee42d63512fb03c201a5b5842997a43e0
GitOrigin-RevId: ecadb30b00e429dca60746fa81a73a6d1efc7479
Change-Id: I27f8210b42bd5a782260c493c1cb03f8c62413ef