[roll] Roll fuchsia [scrutiny] Add v2 component tree as a DataCollection
Constructs a ComponentTree from the set of all v2
components and retains it as a DataCollection.
Also adds a placeholder plugin `verify.map_tree`
which just lists all nodes of the tree in
breadth-first order.

Changes ComponentTreeBuilder's error handling: now,
instead of returning an error immediately if a
ComponentDecl is not available for some child URL,
the builder constructs as much of the tree as
possible. The builder's new return value now
contains both the tree (if any) and a vector of
all errors encountered.

In this CL, Scrutiny just logs each of these errors
as a warning, but a future plugin could operate on
them.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/475138
Original-Revision: b0fdd04a0c3d115db7a07c7c3f4fe8a05fa32686
GitOrigin-RevId: 69dabed36838772ecabbf6cd526be715cb8f9619
Change-Id: I8c13c7773c5f6c86cedc7b6130145ffdd8defa27