[roll] Roll fuchsia [banjo] Make parameters inout if they are tagged with "Mutable".
It turns out that in the async case, some output parameters are passed to
the method formatting input parameters. This means that output parameters
should be handled the same way as input parameters. However, marking an
output parameter as "InOut" does not make much sense, which is why we allow
them to be marked as "Mutable" instead.

Original-Bug: 67708
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474820
Original-Revision: 1c625ddfc35610cf43d545d37247b01085b4c113
GitOrigin-RevId: 04b177dbd417d528e8c73e55bed879d553080a32
Change-Id: Icdce45d98fb087b90f93024626f2f2569a6bcc3f