[roll] Roll fuchsia [fx][qemu] Add an --archive option to the qemu command
Add a --archive options to the qemu command which can be used to
create a compressed tarball with all of the files needed to launch the
VM, instead of actually running the image.  This option can be helpful
when someone wants to take a build over to a remote machine (such as a
Cavium workstation) and run it there instead.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474254
Original-Revision: e7b0ca2ad0fbd428b285078db8a0250b258e73b2
GitOrigin-RevId: d0fc39b9ae0ea0476752f7c452cb92987ae2feaf
Change-Id: I42ed042ef8fd7f413a9864ddb7ac1e1685c6c0b8