[roll] Roll fuchsia [dart] Remove workaround now that type promotion accounts for local boolean variables.
This change removes a workaround that was introduced prior to landing
Dart language feature
https://github.com/dart-lang/language/issues/1274, which allows type
promotion in null safe code to account for local boolean variables.
The workaround ensured that the code would analyze the same regardless
of whether the feature was enabled, allowing for a smoother
transition.  Now that the feature has fully landed, the workaround
isn't needed anymore.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474237
Original-Revision: 3034566cdf9e72591621658e4210f660a8871cbe
GitOrigin-RevId: 6b00c0f686ae207b79dec6d164e44e826ec7f2a4
Change-Id: I6d91af836cd82d199d01b5f75b643a8bede3675e