[roll] Roll fuchsia [i18n] Hermeticize static_icu_data build template
Added explicit references to input files in this build rule.  While
technically we still read the entire directory, in practice the contents
of the directory never changes - it's always the same files.  So it
should be OK to name the only 3 files that are currently there for the
time being.

Original-Bug: 68703
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/475179
Original-Revision: 30f431ef5de643f3f230582d43d3e49f79d77f1a
GitOrigin-RevId: e9db032208186386c196942181fa686d7106e238
Change-Id: Ifd648d0a09a5f04568b82e58d1a27f7b8a6daeab