[roll] Roll fuchsia [harvester] Listen for ZX_CLOCK_STARTED
This introduces a FuchsiaClock class which returns an optional for the
now value of time. We can use this to determine if the system clock has
started or not. We then use this class to supply times to the various
RPC messages that need times. This class has a mechanism for calling a
callback when the clock starts and this is used to send the ClockStarted
RPC to Dockyard.

I haven't added tests yet but wanted to put this up for feedback before
I finish that.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/473862
Original-Revision: bbb471b0e42405eca60839c1fbe36ec9f7fbe301
GitOrigin-RevId: 8af8eb52497e92339018c66dc7615fadb3f20698
Change-Id: I897268a44cba0157bfba204b26221642ef14fe44