[roll] Roll fuchsia [settings] Migrate mic to input device
Migrates the old format of microphone to be in the new format for input
devices. Internally it will now use the input device format when a
client toggles the mic state. The ability to toggle with the old state
should be removed once all clients are migrated to use the new
interface.

Original-Fixed: 67211
Test: fx test -o setui_service_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474113
Original-Revision: 72ba457aa97576bf6b2ff0674015e5236354bc3c
GitOrigin-RevId: 26c858ccfc30a12760cf77589c802a89273b280b
Change-Id: I10bbef2f1d819ed2ea2d650870b0df0a80ecab17