[roll] Roll fuchsia [docs][fx]
Revised the intro to "What just happened?"

It was a little unclear as written.

1. removed the first bullet, which didn't correspond to a separate part
of the example `fx set` command and was redundant with the `board` bullet.
2. Added an intro to the discussion about `base`, `cache`, and
`universe` because these terms are discussed without relating them to
anything in the `fx set` example or anything else above.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474750
Original-Revision: 74a0a932c9a676147b20a93baac44914f6778aa9
GitOrigin-RevId: 8e0302ad87d27637a66975ac7691d5a268a92683
Change-Id: Ib21158e9af920113e4ed2625a28f3dcbf94b8da6