[roll] Roll fuchsia [modular] Serve outgoing dir in FakeComponent after OnCreate
Ensure that the outgoing directory is only served after all
services are added to it, as typically this is done inside OnCreate.
This is correct behavior but effectively a no-op because it happens
in the same async loop frame either way.

More importantly, this enables tests to perform tasks between
creating the component and serving the outgoing directory, useful
for testing racy code that connects to services.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/475743
Original-Revision: 69edb9b1d683c76c183ab0eaa744e55da87c9799
GitOrigin-RevId: 5df3ecaee534a58a542746a70a1bcdffc3fdac31
Change-Id: I5a26c6a806f18768845a18ba679c4563fe924008