[roll] Roll fuchsia [netcfg] Cleanup DHCPv6 state on non-fatal errors
When a non-fatal error was hit when stopping the DHCPv6 client, DHCPv6
state was not properly cleared for an interface.

This change clears DHCPv6 state for an inteface when stopping a DHCPv6
client was successful and when non-fatal errors are hit.

Original-Fixed: 68305

Test: fx test netcfg-tests -- \
        test_stopping_dhcpv6_with_down_lookup_admin

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474753
Original-Revision: 4f54068ac26a53fa3353ccb10db0240ce6a68b21
GitOrigin-RevId: 0c258e792481fe1f8827a2a5a57f1d827e841aa2
Change-Id: Iad65bc926bcffc846d49f3be5dbe3fe43e174638