[roll] Roll fuchsia [kernel][arm64][mmu] use 4K page granules for the bootstrap trampoline
Currently the start.S code is more clever than it probably should be and
compile time decides what size page granules to use to map the large
(1GB/512MB) page around the current code at boot to trampoline to the
high address. It's currently set to use 64KB page granules in the
temporary address space it uses. This is generally fine except 64K pages
are technically optional according to the ARM ARM.

Switch to using 4K pages always which are more or less mandatory, or at
least mandatory to zircon. This avoids any issues booting on cpus that
don't support 64K pages.

This reduces the size of the physical address space the kernel needs to
be physically located in at boot time, but 39 bits is still pretty good,
and it's unlikely we'll encounter a cpu where the kernel is loaded
outside of that range.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/464055
Original-Revision: f12c67c9a363c7a86651d2983c96368c6dadef0e
GitOrigin-RevId: 848aa72990f4b670d85efa8432a4fd5d81767b4d
Change-Id: Idf77613197a38593011838cc6f54f6ba6a5d412c