[roll] Roll fuchsia [kernel] Unittests of kernel stack depth handling
Add a unittest that checks that we can receive interrupts and handle
mp_sync_exec calls when thread context has used up half of the kernel
stack.

The first test is simple and probablistic - it allocates a large
on-stack buffer and waits a bit for any (potential) interrupts. If an
interrupt were to take > 1/2 the kernel stack to handle, we would see an
architecture-specific stack overrun failure.

The second test is deterministic and ensures we can run an mp_sync_exec
handler (that signals an event) while half the kernel stack is used from
thread context by a buffer.

Original-Bug: 64808 Panic on boot on the kasan bringup builder
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/461084
Original-Revision: d6c3fbc9bfa3812a9aee187e1664d5496d66f78e
GitOrigin-RevId: 5292fd2d8f0e194183c7bedbc38a83f9b8093f5a
Change-Id: I69f873aa0292bc283097d4d00150e27d69d48cd2