[roll] Roll fuchsia [console-launcher] add /hub
The hub directory was removed from the serial shell some time ago,
because due to a CFv2 migration requests into this directory would hang
on bringup and this broke the shell.

Later the component framework was configured to disable the fuchsia-pkg
resolver on bringup, which solved this issue, so the directory can be
safely reintroduced.

Original-Fixed: 50805
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/476258
Original-Revision: 3d9a9fcc1526c27f4ffc2044082efd857fb3323a
GitOrigin-RevId: f40df102cf8aea5a946eaa6c6835f5e52e5118bf
Change-Id: I69c886d00168b7628e587c66ac065fe31fbb9167