[roll] Roll fuchsia [system-update-committer] implement wait_and_reboot
This change adds reboot functionality to the system-update-committer.
Generally, if we fail to put the boot metadata in a happy state, we'll
reboot. Exception: if it's a blobfs verify failure, and the config says
to ignore, we won't reboot.

Note: we can't yet integration test the exception because currently the
blobfs verification will always pass.

Original-Fixed: 64600
Test: fx test -o system-update-committer-tests
Test: fx test -o system-update-committer-integration-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/475560
Original-Revision: 4acf3c4c37be79e52029cd2c3cc52de5b20608dd
GitOrigin-RevId: e6d955841d59838f9fb0efa5a27b9ee7081d0db9
Change-Id: I908aad2c0efa2a724616f1dd819a6f42b417d1ed