[roll] Roll fuchsia [blobfs] Cleanup in blob test utils
Remove the Merkle tree validation from GenerateBlob. The digest library
is responsible for testing the Merkle tree generation so it doesn't need
to be done here. This removes all of the ASSERT_* calls from
GenerateBlob and GenerateRandomBlob allowing them to return the
generated BlobInfo.

Replace The ASSERT_GT in GenerateRealisticBlob with a ZX_ASSERT_MSG
because it's checking an invariant of the test util and not an invariant
of the code being tested.  This allows GenerateRealisticBlob to return
BlobInfo as well.

Stop returning a status from CreateMerkleTree and instead ZX_ASSERT
inside of the method.  Failing to generate the Merkle tree is a bug in
the test and not a bug in the code under test.

Remove the merkle and size_merkle member variables from BlobInfo and
remove the Generate*Blob overloads taking a BlobLayoutFormat.  The
overloads were added to support the new Merkle tree format in blobfs but
very few tests actually use the merkle tree member variables so the
overloads were making the common case more complicated.  If tests need
the Merkle tree or it's size they can call CreateMerkleTree.

Change the type of BlobInfo::data to use uint8_t instead char to match
the rest of the blobfs code.

Remove several unnecessary std::move calls.

Update some function params to better match how the params are used.

Original-Fixed: 63295
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/470532
Original-Revision: 08ff9d3b731bea860799e6011db9bcbb635ad939
GitOrigin-RevId: fc2f26367bdc69fe76ef599c2ab76eb34af757ec
Change-Id: I56eabd3bc4ca4252e473f3a9932f2eda55d5fa4a