[roll] Roll fuchsia [stress_tests] Remove InstanceUnderTest for actors
Environment now stores the Actors used in the stress test.
Now when an Actor fails, the environment has a `reset()` method
that is called.

The environment can reset the world and update the actors with
new parameters, hence eliminating the need for InstanceUnderTest.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/473898
Original-Revision: 39f6f4fec7989d007469883a49571b68cfa0abda
GitOrigin-RevId: 00edfbdd056360c63d0a38ec2dbedf6c64e5a4ce
Change-Id: I0ba3bed4345e2150d46271e3cc7142af47035770