[roll] Roll fuchsia [wlan][policy] Allow SME layer to complete its re-connect attempts
When it detects a disassociation, the SME layer automatically attempts
to reconnect. This reconnect is visible within the `connecting_to`
field of the SME status response.

Prior to this change, the Policy layer ignored the `connecting_to` field
and performed its own reconnect logic. When the SME layer was already
reconnecting, this interrupted the reconnect and resulted in a longer
delay to get reconnected.

Original-Bug: 65611
Test: fx test wlancfg-tests
Multiply: wlancfg-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/476519
Original-Revision: a45e61c7b421a1dc2e10a3a5e14584cd39225d13
GitOrigin-RevId: c8fd20a079389a867895331da805ba887cdff72d
Change-Id: If6455b4b8fd34d6b664ef99c2c3a3a5a91405798