[roll] Roll fuchsia [dhcpd] consistent recovery from stale configs
Previously, the expired client configs will still be in the server
cache, but they are not allocated in the address pool. This will
cause panic when netcfg tries to call clear_leases(). Added a test
to catch this case in the future.

Test: fx test dhcp-tests
Original-Fixed: 62989

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/472101
Original-Revision: 0ffaf50fc542cdc2ee5c77685fe14785ccfa1e88
GitOrigin-RevId: 4dc31239c3fcdbc23980ae933baf04b85e546267
Change-Id: I64dd100b3b7788b07b369548198326a38d3bad70