[roll] Roll fuchsia [brcmfmac] Remove ASSERT in brcmf_notify_start_auth()
This function used to ASSERT when seeing an AUTH_START event while the
iface is not in CONNECTING state. This scenario is verified to be
expected by ACTS test, when an WPA3 authentication is interrupted by an
deauth_req from SME before firwmare sends up the START_AUTH.

So the ASSERT should be remove and the log level should also be changed.

Test: None, no functional change.
Original-Bug: 68540
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/476937
Original-Revision: 68d479bfbee54701f41f1f2735983fa5c1845471
GitOrigin-RevId: 6622c20155e823e2d6cc37fdabf17251be69aad7
Change-Id: I61088e106335125eba3ee08d48d6d66b38527833