[roll] Roll fuchsia [Bind] Encode the symbol table for new bytecode format
Add headers and the symbol table to the new bytecode
encoding. This CL modifies the Condition enum so that it
stores Symbol values instead of u32.

For additional clean up, this CL deletes c_generation.rs
instead of updating the Condition code in the file since
the file is unused.

Original-Bug: 67440
Test: Added unit tests in encode_bind_program_v2
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/472842
Original-Revision: 9bfb82980fc105092dfc07ea78edf520555ea761
GitOrigin-RevId: d9ef1bf4e296eec015f4835cf7c2cc44d8ff2bd9
Change-Id: I85476feed3f4a4063958c3068f2cfd9354ec79f7