[roll] Roll fuchsia [wlan] FIDL MLME: rename *Codes -> *Code, move ReasonCode to ieee80211
This is a functionally no-op change that renames the *Codes (e.g.
JoinResultCodes, AuthenticateResultCodes) to the equivalent Code (e.g.
JoinResultCode, AuthenticateResultCode), better matching the MLME
specification in IEEE 802.11-2016.  It also removes the ReasonCode
enumeration within the MLME FIDL and moves to using the ReasonCode
enumeration as defined in the ieee80211 FIDL.  In the process, some
usages of the ReasonCode that were previously typed as uint16 are now
explicitly typed as ReasonCode, for better semantics and type safety.

Original-Bug: 29277
Original-Bug: 55667
Test: fx test wlan-sim-env-test wlan-sim-fake-ap-test brcmfmac-sim-test
      wlanif_unittest wlan_tests wlan-mlme-tests wlan-rsn-tests
      wlan-sme-tests wlanstack-tests
      Booted and connected to WLAN on Astro

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/476697
Original-Revision: 1517bafcca534089b649b0ce8257c09f3d96fcb2
GitOrigin-RevId: a5e7d6e995b78ec299b4f015b9a80742ae503134
Change-Id: Ib3e09cccb53c14817ea91d3234275a4e162a6067