[roll] Roll third_party/boringssl [BUILD.gn] Export explicit static targets
Drivers on Fuchsia are expected to statically link all dependencies, so
the default targets (which for Fuchsia targets are shared libraries)
aren't what we want for drivers that need to link boringssl.

So, we define additional libcrypto and libssl targets to cover the build
matrix, and carefully make sure the output name of the default library
type is preserved to avoid changing behavior for existing usage of the
:boringssl target (since otherwise that seems to trip up Rust targets on
host with `non_rust_deps` at link time).

Then, we define an additional `:boringssl-static` target which can be
used to reliably pull in the statically-linked library for use in
drivers.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/boringssl/+/476857
Original-Revision: 0e1398c94ad9791b5de876438af8c377408c1d9b
GitOrigin-RevId: c7f99ab47664eba090fc6c527fed5dae02faa509
Change-Id: I8a369a8bc5c934123eff1198de4d8a799fe9243b