[roll] Roll fuchsia [build] Replace profile_list with profile_source_files and profile_source_files_list_files
This gives a clearer name and documentation to the GN build argument
for providing a prepopulated file for selective instrumentation, and
makes it a list of such files.  It also adds another build argument
that takes the list of source files (as GN paths) directly and
populates a file to feed to the compiler.

Original-Bug: 68518
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/477513
Original-Revision: 129b77bea9be1e13aa0367a4a36276f6bce40164
GitOrigin-RevId: 9fc5685cf2eb3702fa7a1acaa4d317b68b67d383
Change-Id: I291e25fee6d711b91f156b629db245a3f602cd2a