[roll] Roll fuchsia [kernel][arm64] Sync executable ranges prior to mapping
This moves the icache sync logic from the page fault handler into the
arch mmu code. In the process the icache sync is also moved to happen
before a mapping is installed. This removes a race condition where
another thread in the process could use the unsynced mapping.

Performing the icache sync in the mmu code, instead of the VmMapping,
corrects an existing case where sync is not done when mapping via
VmMapping::MapRange, and prevents any future chances of forgetting to
sync.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/470698
Original-Revision: 49d0e3fd006b237a67f414cae3edafcc2752e593
GitOrigin-RevId: 8f62f2147ee7245d6111675681b7fad97122b28b
Change-Id: Ifb6e951a17933c2e2dd0e4c57ed0063069295253