[roll] Roll fuchsia [licenses] Add new zstd cases
We're cleaning up to use a verbatim github repo mirror for zstd.
The upstream project is dual-licensed and so its sources contain
both a BSD "LICENSE" file, which applies to our use, and a GPLv2
"COPYING" file, which does not apply to us.  Since we track an
exact mirror of the upstream repo, it's not an option to remove
the files we don't use from the tree altogether.

When the transition is complete, the existing entry for the
zircon/third_party/ulib/zstd directory will be removed as that
directory will no longer exist.

Original-Bug: 68624
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/477738
Original-Revision: df9ac37795db03661b24ecf86ce4fdd0e48e2f2b
GitOrigin-RevId: e81bdc28c2aa5c24dd9708338f97bf24f61f5383
Change-Id: I25613c14ea95e0a266607b55e3644e231930a192