[roll] Roll fuchsia [netstack3] Stricter typing for IP address
In `net-types`, introduce a new `Ipv6SourceAddr` type, which
represents the source address on an IPv6 packet - such addresses
can only be unspecified or unicast, but not multicast.

Modify the IP layer API so that, when an inbound IPv6 packet is
received,
- if its source address is multicast, it is silently dropped
- otherwise, its source address is stored as `Ipv6SourceAddr`

This allows consumers of inbound IPv6 packets to rely on stricter
type bounds on the source addresses of those packets.

Also modify the `icmp` module's functions to accept `SpecifiedAddr`s
where appropriate, and push validation to the caller (the `ip`
module).

Test: fx test netstack3_core_lib_test -- \
      test_drop_non_unicast_ipv6_source

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/476508
Original-Revision: 857bc8260072e5717ec957b7d42542c34770ec82
GitOrigin-RevId: 2b4dd21507891b86bde12ae94fe862aafdad35fd
Change-Id: Ie0aa0732bfe5d138a11ec7fa92a15ddea330761c