[roll] Roll fuchsia [netstack] Implement error interface on pointer
Implementing interfaces on value types causes the interface to be
implemented by both the value type and the pointer type of the
implementer. This is a footgun because it makes it harder to write
correct type assertions; it requires the assertion to check for both the
pointer type and the value type.

Since interface boxing always causes the boxed object to escape, the
performance implications of implementing interfaces on pointers is nil,
and the result is better type safety.

This changes TcpIpError's error implementation to the pointer type.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/477879
Original-Revision: 07ec2f4d0d6fe0aa3d7af662cdaf69a86b811b7c
GitOrigin-RevId: e20774a01566769723b104b548cb4db997503736
Change-Id: I5a1ae738a1b80d33da44d42f8e43b4ee0227d4ee