[roll] Roll fuchsia [brcmfmac] Refactor async_dispatcher usage in Device classes
This commit refactors the use of async_dispatcher in Device classes.  In
hardware-associated classes (PcieDevice, SdioDevice), the
async_dispatcher is provided by an owned async::Loop.  In the SimDevice,
the async_dispatcher is now refactored to use the dispatcher provided by
the wlan::simulation::Environment.

By standardizing all Device classes to provide task dispatch through an
async_dispatcher, we are able to refactor the brcmf_bus_ops interface to
remove the .set_sim_timer and .cancel_sim_timer entry points, which
allows us to simplify the Timer interface to remove its dependency on
the bus.

We are also able now to remove the Device::Init() entry point which was
previously a layering violation (required to be called by subclasses),
and to simplify the lifetime semantics of the Device and remove the
explicit Start() and Stop() entry points.

Test: fx test brcmfmac-test brcmfmac-sdio-test brcmfmac-sim-test
      Booted and connect to WLAN on Astro

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474824
Original-Revision: 0c2b029bc1b282316c67fb1446c311f1f1d2afb0
GitOrigin-RevId: 2dcc76b76e8e9003dd00525ced640d630eedc666
Change-Id: If5761631af9298f7b1d708054a184b65ef75abfa