[roll] Roll fuchsia [brcmfmac] Fix workqueue test flakiness
The workqueue tests were susceptible to issues with spurious wakeups in
sync completions. Since a sync completion can wake up without being
signaled we also need to check some kind of condition to determine if we
should keep waiting or not.

Add a small support class that combines the sync completion with a flag
indicating if the completion was signaled or not.

BUG: 68548
Test: Ran test 100s of times with no issues
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/477601
Original-Revision: 7a76d9a35571c32e87f33328ebd1da610fb04cac
GitOrigin-RevId: 2e55f3d976ff808d8b2f147faae410ee8c57375d
Change-Id: Iba82255eb41cf3ad88c7ad3c72e6a8f2c8b42c18