[roll] Roll fuchsia [build] Remove archive targets from uefi-disk target deps
These files are not actually inputs to uefi-disk action target,
but their targets are erroneously in its deps.  Since that target
is in the default graph for x64 builds, building the archives is
in the default graph for no reason.  This change fixes that.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/477957
Original-Revision: 2f584c4a62374f37361ac04875e60b5459fcc3b5
GitOrigin-RevId: 38401e9406144ba2438abcde07421bb1a3a516d1
Change-Id: I52bb26fa4fbd8ea31c4caba171e395a10d980fe7