[roll] Roll fuchsia [fx emu][gn sdk] Fix flaky tests in femu-test.sh
The previous checks on sudo commands was flaky. The test was running
the entire script and trying to look for the last .N file to check
for a command it was expecting. Also, the start-unsecure-internet.sh
script does cat /proc/sys/net/ipv4/ip_forward and this file may not
exist or be set to 0 on some Linux machines, and would also fail the
test.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/478529
Original-Revision: 8f363af5df89de785cb721a7eae26edcb7c07154
GitOrigin-RevId: 0759263e4bda1fdc10905f5d904d09d90b523236
Change-Id: Ic6cbd063f886ef0ed28dd168d2e23253851b5263