[roll] Roll infra/recipes [autocorrelator] Improve property ergonomics
Instead of requiring that the full builder name is specified, only
require that the CI bucket is specified. The project will always match
the project of the current build, and the builder name will usually
match the builder name of the current build, unless if overridden.

So in most cases, the only value that needs to be set is ci_bucket,
unless the CI builder does not match the name of the try builder, which
can happen in some special cases.

Note this is technically a hard-transition that will disable running
autocorrelator in CI for builds since they don't set the ci_bucket
property, but since there is only one FYI builder at this time running
autocorrelator, we won't bother with a soft-transition.

Original-Bug: 67661
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/478597
Original-Revision: 81d14dbb430afac39d37aa8a44c432d43441dd7a
GitOrigin-RevId: 2a562bfce0452ccf2f64409b9d2f0d41ff076718
Change-Id: Ib3b3f0dcc44dbfe63323f700c30f87e388ffbe2b