[roll] Roll fuchsia [netstack] Require link resolution in test
https://github.com/google/gvisor/commit/daeb06d removes link address
resolvers for interfaces that report not needing them. This causes the
DHCP test to fail on links that do not support link address resolution.

This marks the test interface as requiring link address resolution and
allows the test to pass before and after the upstream change. This also
replaces loopback with the test endpoint where loopback was used.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/478539
Original-Revision: 955a29372428d33590111874e60c8b51df9c30af
GitOrigin-RevId: e4251e50d6fd484abb35ad463eed053d9ea1eeff
Change-Id: I9c1f09ed0efc18fe01658b78167330299699a3e3