[roll] Roll fuchsia [memalloc] Allow iteration over available ranges.
Add support for iterating over the ranges currently tracked in the
allocator.

We'll shortly need this for ensuring that physical memory tracked by the
allocator has been mapped into the current address space, and a little
later on when we want to release all the unused memory in the allocator.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/477921
Original-Revision: 9d969a12018329f89ceb12e10cb7210050e15e48
GitOrigin-RevId: 4e7adad8d6950b3d41aab64a514849ddd9044616
Change-Id: I866d4a92b4a493e625c5d66135b706edb3208506