[roll] Roll fuchsia [make-fuchsia-vol] Fix bugs in and usage of optimalBlockAlign
optimalBlockAlign wouldn't always return the correct values. This is
because it would incorrectly chop off the end of a partition if the
partition size was not divisible by the logical block size of the disk.
It would also incorrectly add a block to partitions that were the
logical block size of the disk.

While we're here, fix the calculation of the size of the FVM partition,
where a separate off-by-one error would result in a partition that was 1
block shorter than intended.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/477097
Original-Revision: d7deddf07e5e2199f71b112cb09caac18f4c416e
GitOrigin-RevId: 195e4e7a6bc99bb3f208251f81cb296135606750
Change-Id: Ie610c87126902914ebe9f4227f944f344daa9d8e