[roll] Roll fuchsia [netstack/dhcp] Remove flake in tests
Commit 52021143eea968c388fdf366d843b7aa5ad371c9 introduced a change to
asynchronously deliver network packets to the network dispatcher in
tests. This removed the ability for tests to reliably synchronize with
packet delivery.

This change adds a callback to the test endpoint that is called after a
packet is successfully delivered to the network dispatcher that tests
may use to synchronize with packet delivery.

Test: fx test netstack-dhcp-gotests -- -test.count 100
MULTIPLY: netstack-dhcp-gotests: 100

Original-Fixed: 68050
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/479177
Original-Revision: 0ccefe8f0caced00a05e4282c743d2d383715df9
GitOrigin-RevId: 9bad2e7b5ebc84da0b9495539bdf56addd8dd741
Change-Id: I2910b6af7086a5d70593d60eadee2577056404be