[roll] Roll infra/recipes [hooks] Run Black on changed files pre-commit
This should hopefully prevent more formatting breakages that get through
CQ because the reviewer doesn't notice a Tricium comment. It should also
make it less common that a commit author pushes and starts CQ, only to
realize that a file is malformatted, requiring another push and CQ
attempt.

We'll only run Black on files that have been changed since master, to
make the check faster and make it less likely that a commit that breaks
formatting will prevent everyone else from committing.

This adds around 0.25s to the pre-commit runtime when I tested it
locally with 17 modified Python files, which I expect is on the higher
end in terms of modified file count.

If there are formatting errors, the script will block committing and
output something like the following:

  Files not properly formatted. Run 'black .' and then 'git add'.
  would reformat fuchsia/recipes/recipe_modules/artifacts/examples/full.py
  would reformat fuchsia/recipes/recipe_modules/emu/examples/full.py
  would reformat fuchsia/recipes/recipes/linux_sysroot.py
  would reformat fuchsia/recipes/recipes/kythe.py
  Oh no! 💥 💔 💥
  4 files would be reformatted, 19 files would be left unchanged.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/479199
Original-Revision: 2410dddf005ffd57f38ac57841fcc4331775316a
GitOrigin-RevId: 875b8297ea6467aba3aa170f2cf8efa070cd53d2
Change-Id: I3d69058964f22ef014ee9a0e9d7a670851b9f650