[roll] Roll fuchsia [bind] Refactor tests in encode_bind_program_v2
Refactored the tests in encode_bind_program_v2 so
that it's easier to read and write them, especially
when the tests become more complex in the future.

This CL introduce constants for the number of bytes
in instructions so that the tests can use them to
calculated the expected results.

It also introduces helper methods to verify the
instructions and the headers.

Original-Bug: 67440
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/478638
Original-Revision: bf5c3308dca447d710e1442364a420614298ba2a
GitOrigin-RevId: 9ce12df62a3af9f94b2a9e54e8d5bafda53b8160
Change-Id: Id0ccba4fb2ea96c87e5fd8c4b1fa6980057c9416