[roll] Roll infra/recipes [zbi_test] Convert to proto properties
I'm planning to add an integer max_attempts_per_test to this recipe, but
bbagent historically has decoded integer properties as floats rather
than ints, which leads to some weird behavior. To avoid that issue,
convert to proto properties so that properties that expect an integer
always get decoded as integers rather than floats.

This is technically not backwards compatible because protos don't let
you specify a default value for a field that's not the zero value for
the field's type. So `use_goma` must now default to false and
`test_pool` must default to an empty string. This shouldn't be an issue
because we explicitly set these properties on all builders, so the
defaults were not load-bearing.

Also stop wrapping the `checkout()` call in infra_steps=True, because
the checkout recipe module internally marks steps as infra steps where
appropriate.

Original-Bug: 68783
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/479198
Original-Revision: 7969f02f12e9e01dac4f3d3e1bb57d8e90f507d6
GitOrigin-RevId: 4ae93e459a7c7916ec4d51d3ea0b3a15642b05f6
Change-Id: Ifab183bf461e13f2260f302b5fbce66360ca98e2