[roll] Roll infra/recipes [zbi_test] Add a max_attempts_per_test property
This will be passed into swarming_retry as the `max_attempts` parameter
to control how many times we retry test tasks in case of failures. We'll
use this to lower the number of attempts in CI to 1, while leaving the
number of attempts in CQ at 2.

Original-Bug: 68783
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/478900
Original-Revision: 1964548181e5a4c18355819cbf9182cd91a422d4
GitOrigin-RevId: 958bce154d68d2d6881601bc57c870956a532885
Change-Id: I8166742ab1ff13db5b55ec39a4e83b0a38b8b3e7