[roll] Roll fuchsia [fidl][llcpp] Fail subsequent calls after a client has been unbound
In Icd592da6f0e9daaed9d46d94651d6055a9f06133, the lifetime of the
channel was decoupled from the bindings on the client side (see that CL
for the background and context). That introduced a silent regression
where it was possible to make calls on an unbound client that would
seemingly succeed, except that nobody would be handling the responses
anymore.

This CL fixes that regression by still maintaining the decoupled
lifetime, but getting the channel from the binding instead of the
ChannelRefTracker when making calls.

In short, when a user |Unbind|s a client, the channel still stays alive
in the client through the ref tracker, but one cannot make any further
FIDL method calls. The client enters a "half-closed" state, ready for
eventual cleanup via its destructor.

Test: Added one in `fx test -vo fidl-walker-tests`
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/478624
Original-Revision: 023df8339c8a2ef9b338515bf722781754e1f25a
GitOrigin-RevId: 69f319248e094fe79cdfa65ec09af1a1a9fa6615
Change-Id: Iad785a254e4fb112d9a2aa4f2a9488de8206ba0d