[roll] Roll infra/recipes [display_util] Improve build presentation
In order for the tree closer to be able to filter out SDK builds that
turn purple due to subbuild checkout failures, we need to propagate
subbuild summaries to the parent build's summary_markdown.

To do this in a way that wasn't super ugly, I had to introduce some
duplication of logic that was previously centralized in the `_display()`
function that was shared by both `display_builds()` and
`display_tasks()`. I think the duplication is worth it for simplified
and more readable code, and it makes it much easier to add new logic
that makes sense only for Buildbucket builds or only for swarming tasks
(like summary markdown manipulation - Swarming tasks don't expose text
summaries).

Also:
- Don't automatically wrap subbuild summaries in backticks. This only
  works if the summary is known to contain code-like text (e.g. stdout
  from some process) which is generally not the case here.
- Use markdown to link to the subbuild from the parent build's summary.
- Use `pluralize()` to indicate how many builds/tasks failed while
  maintaining grammatical correctness.

Original-Bug: 58963
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/479033
Original-Revision: f3c915969de768aeb731c01d622a9663b9cab9dc
GitOrigin-RevId: e6dda9cd463821aa17c78d6ce569fb3d61a4b54c
Change-Id: Ice1efb931dd496fa600e8e2487be457c42f62577