[roll] Roll fuchsia [kernel] Add support for many relocs in gen-kaslr-fixups.sh
This CL adds a workaround for ancient versions of awk that did not
support having many references to the same string in memory.

The gen-kaslr-fixups.sh script executes an awk program that, for error
reporting purposes, kept a dictionary of relocations to section
mappings. Builds that have too many relocations in the same section
(>65k) would overflow the reference counting in old versions of awk,
leading to segmentation faults.

The fix just makes sure that each entry in this map is a unique string,
by concatenating the section name with the relocation address.

Original-Bug: 68230
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/478019
Original-Revision: 8fd974f68b1bb61e15e41b0bcd570c232f087089
GitOrigin-RevId: 930a01de0f6cd72ddf67484cf119c189d4ca4c17
Change-Id: I035ccebf16234234e81c9cdbe2cf834c89a98a14