[roll] Roll fuchsia [bt][profile-test-server] Clean up component launching in tests
Context: The unit tests in the PTS launch A2DP with source enabled.
AVRCP-Target is launched as a byproduct of this. Because AVRCP-TG
requires extra capabilities (PeerManager + ComponentLifecycle) which are
not present in the sandbox, launching fails. The extra latency
associated with the failed launch has exposed flakes in the unit
tests, particularly on AEMU devices. These flakes are seen in the
ordering of component termination events. Currently, the event streams
for each launched component are polled sequentially, in the order that
they were launched in. For the most part, this flake was never seen
previously because there was no extra overhead in launching the profiles
since bt-avrcp-target was never launched as a side effect. However, the
general assumption that launched components would finish launching in
the order they are defined in and therefore generate termination events
in that same order is unsafe. Furthermore, the way in which we are
forcing component termination has introduced flakes. Currently, we rely
on the A2DP/AVRCP profiles to terminate when references to the
registered Search & Advertisement are dropped. While this does result in
component termination, there are weak guarantees around the timing of
the termination.

In order to eliminate these flakes we do several things:
1) Disable AVRCP-Target in the PTS unit tests. The tests are simply
verifying that the profile-under-test gets launched correctly. AVRCP-TG
is not required for this.
2) Separate functionality of the test into two separate tests. One to
test launching + terminating a component and another to test launching
multiple profiles concurrently.
3) Force component termination by using the underlying App, as opposed
to indirectly via dropping the Profile handles.

To make the tests more robust, particularly when run in parallel, we:
1) Randomize nested environment labeling by using salt.
2) Use unique peer IDs in each test to clarify that these are distinct
peers.

Original-Bug: 67526
Test: Re-ran fx test bt-profile-test-server-tests

Multiply: bt-profile-test-server-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/475801
Original-Revision: d9b441f3f7862ad03ea5776b973ee3ec87e13cad
GitOrigin-RevId: 17910d764b91d76d2b5fe6d94077bf17d2c2a65f
Change-Id: I641794bba968fc4e0aae2d4580ba71c63e873269