[roll] Roll fuchsia [kernel][vm] Optionally free page tables and rename API
Extends the ArchVmAspace::FreeUnaccessed method to take an optional
parameter to control whether it just harvests accessed information, or
also reclaims the page tables. As part of this the method is renamed
to HarvestNonTerminalAccessed, since it is not always freeing.

Optional freeing allows for regularly harvesting the accessed bits
in preparation for reclamation, but only actually reclaiming when
desired, such as when actually under memory pressure.

Original-Bug: 34565

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/477923
Original-Revision: f993147d5f98605c2094f901f34191ce7a2330e3
GitOrigin-RevId: 7fa9bb4878a259e128fbb64cff86785ea4197e4f
Change-Id: I8dea2186b0ff5f1e7350fa58be6b2e420270257b