[roll] Roll fuchsia [events] Remove BlockingEventSource
With this change, there is only one kind of EventSource and individual
events are either sync or async. All affected tests are updated to use
EventSource. For OpaqueTest, there is an implicit static EventStream
that listens for Resolved and is dropped when the EventSource is
dropped.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474625
Original-Revision: 608ef1f1d2b2951dac401e72f0a1123c741815fd
GitOrigin-RevId: fc15e5c5c901abee7801b27c10480c62d5c04d9c
Change-Id: I214f0048591b0cec6207f49d8e303e10fe9f2ed2