[roll] Roll fuchsia [session_manager] Move session_manager into core.cml
This change moves session_manager to be a child of core.cml.

This allows us to delete the component manager instance that was unique
to session manager.

The change follows the same pattern as was used for appmgr to make sure
that the system does not go down when session manager is not available.

Original-Fixed: 64080

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/471639
Original-Revision: b32aa915b4c937a04da2e76944fa1b697f948643
GitOrigin-RevId: 5e3da06a35b1f038d21751ab765063807b1e6afe
Change-Id: If9b59ff1545c9aff91cdea8772a1a5be3a70c535