[roll] Roll fuchsia [fvm] Fix partition size limiting.
The device path used by fshost for FVM was incorrect so size limits
were not being applied.

Add more logging to FVM failures.

Adds additional logging for failed partition lookups, especially around
partition size limits. There was a report of a failure and it's not
clear why FVM is issuing the error.

Separate out the FshostIntegrationTest fixture implementation into a .cc
file and move the waiting code from the recovery-test to there so it can
be shared with the new test. Rename the test fixture file to match the
class name.

Reorganize the declarations of the fshost config constants so they are
easier to follow and the documentation can be longer.

Original-Bug: 64055

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/475442
Original-Revision: 989fe8bfcb1d8aae973df9b7aaccbf56536fa6e0
GitOrigin-RevId: 96674de87805e9bb1ee03b7f5d6504fa4339bb84
Change-Id: Ie6bb9383d7a7eaaacb9d6fdce5db8303cfce1926