[roll] Roll fuchsia [wlan][policy] Connect to WPA3 with WPA config
This change allows network configs saved as WPA to be used to connect to
observed WPA3 networks when auto connecting.

This change does not allow networks saved as WPA3 to be used any
differently.

Security type is not used as a determining factor in
network selection, so if we see two APs with the same SSID and WPA and
WPA3 security, both will match with a config saved as WPA and the actual
security type will not be used to select a network.

Original-Bug: 472909
Test: Added a unit test and modified existing tests:
    fx test wlancfg-tests
  Manually tested that a network saved through the donut tool as
  WPA will trigger a connect to a WPA3 network.
Multiply: wlancfg-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/480009
Original-Revision: ce39c283784e08dfc624e1223fbb5ec7c4a515b9
GitOrigin-RevId: 4545c15388dd75cca5eb301553cdf8af93e3ab94
Change-Id: Ib82049faec3996c33d8ea2566f60440e90f2fe63