[roll] Roll fuchsia [fx] Create a root .fx/ dir for caching artifacts between runs
For now, this is only used for caching the new fx-set Go executable.
Eventually it's probably a good place to centralize the information
that's currently in one-off files like .fx-build-dir.

It's already in .gitignore so it shouldn't be noticeable to any users.

Original-Bug: 68465
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/480048
Original-Revision: 4715bb6ad267c8d0410aaf4b5a1de7be2969fc79
GitOrigin-RevId: 8cbc5ca2c5a74f374094fbdf04ed02a456cea0dc
Change-Id: I75ccf8465e692341be2fac0be2057d8999d41b24