[roll] Roll fuchsia [pkg-cache] Implement fuchsia.pkg.PackageCache/Get
This change replaces the existing stub handler for the
fuchsia.pkg.PackageCache/Get API with a functional implementation.

While it is currently implemented using pkgfs, a future change can
replace its implementation with one based on blobfs without needing to
change the FIDL contract.

Most of the logic in pkg-cache to serve this API (and the NeededBlobs
protocol) was in prior CLs, though this is the first one to change the
public API of pkg-cache. So, this change also starts on pkg-cache
integration tests that cache packages through pkg-cache instead of
pkgfs.

Original-Fixed: 64624
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/475677
Original-Revision: d08b20b6af88af5341ff300d55a802de20b449cd
GitOrigin-RevId: 18b09b7edb676a8cbfaa2ec7882c6cffbe2dcf9d
Change-Id: I1483e0f1ba77dce71c17cba144bdfbd8421b7351