[roll] Roll fuchsia [bind] Encode jump statements
Encode jump statements in the new bytecode format.
Since the jump statements require calculating the
jump offsets, a map that contains the label
ID, the label index and the jump instructions
is added to the Encoder struct.

The map is updated when the bind program encounters
a jump instruction or label. Once the Encoder
finished encoding all of the instructions, it iterates
through the map and update all of the encoded jump
instructions with the calculated jump offsets.

Original-Bug: 67440
Tests: Added unit tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/479117
Original-Revision: 09fd7ad38c6ed97b539c329d8392f9b038a05d5e
GitOrigin-RevId: 90feae1077d9cb7360a055762ec8a7d0636cd377
Change-Id: Ic770c1eb983a1ed4626f6f0367c10a763766c095