[roll] Roll fuchsia [cml][runner] Remove `from` property of RunnerDecl
The `from` property of the RunnerDecl was initially added to
add another round of routing, so that a protocol could be routed
to a `runner` capability, which is then routed to an environment.

This indirection is unnecessary and routing can always be performed
from the source component in the form of a `runner` capability.

Original-Bug: 52195
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/479197
Original-Revision: 52fce0aef83873b4c0fa4a507b2c90572b90f722
GitOrigin-RevId: 3b7720efa8ddfd521c51c23ac4f447ec1b48ece9
Change-Id: I7bbf361c42ac75c0261de367219a4c4134496ba6