[roll] Roll fuchsia [sshutil] Don't hold lock while reconnecting
Previously, `Client.Reconnect()` would hold the client lock while
reconnecting. This would cause all the other `Client` methods to block
until we stopped attempting to reconnect (which might be never).  Most
significantly would be `Client.Close()`, which typically gets called
with a `defer client.Close()` while unwinding because of an error.

This fixes this issue by not holding the lock while trying to connect to
the target device. This has the unfortunate side effect that it's
possible to have multiple reconnection attempts in flight, so this patch
detects this case, and just closes the excess connections.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/480148
Original-Revision: aec9d38a816ebf685ea9d38a7f8b25d6b071fbc7
GitOrigin-RevId: 6f44441a911e0583c3e0751c4d611cbb53057ba7
Change-Id: Id49217958a0fa7abb7ffb024c04176b6b35b8156