[roll] Roll fuchsia [zbi] Update ZBI spec for modern booting design
This changes the wording to remove the "complete ZBI" term
and instead refer only to a "bootable ZBI".  The definition
is simplified to require only the kernel item, with the other
items required left to the particular kernel's needs.

The zbi tool is updated to replace --complete with --bootable
and its check simplified to match the new definition.

Also add the new type ZBI_TYPE_STORAGE_KERNEL, which will be used
in the new "physboot" booting design for Zircon.

Modernize the formatting of the header file.

Original-Bug: 57245, 68585
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/480364
Original-Revision: 159be7cd80ae662a3461426f4c58dbd02e892daf
GitOrigin-RevId: 29d0fddf46f8c3058b20ebce7bb3b4e38cf037ca
Change-Id: Ic63519d59712d6956385bd8ff04f755c4cef7386