[roll] Roll fuchsia [components] Modify directory offer to use "as"
It looks like there was a typo in the example on how to define offers. In the "directory" example, it
lists "directory" twice, but it appears that this section was trying to explain how offers could be
renamed with "as", so I switched to using that.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/462495
Original-Revision: f046cc15a343489ed3be40c4837917dc11505887
GitOrigin-RevId: 61db915f4862c6cf46174f1fafba6009a2bb1535
Change-Id: I0464d0f45b08af93d78f9b7c1571fc829cc2095f