[roll] Roll fuchsia [fidl][fidlconv] Add working proof of concept for ConvertingTreeVisitor.
This change serves as a minimal implementation of the chosen design for
the fidlconv converter.  It contains converters for (most) types, struct
declarations, struct members, and name-type ordering (ie, "uint8 FOO" ->
"FOO uint8").  More converters (and tests for them!) will be added in
future changes.

The basic idea of the design is to encapsulate every type of transform
specified by FTP-050 into its own "Conversion" class.  Some of these
transforms can be nested within other transforms (example: "alias foo =
array<handle?>:5" contains a transform of the "handle" type definition
within the transform of the "array" type definition).  Because of this
property, the ConvertingTreeVisitor stores Conversions on a stack,
taking the output of each previous transform and adding it as a "child"
to its parent.  When the stack is exhausted, the final result is written
to an output string containing the text of the fully converted file.

Test: fidldev test fidlc
Original-Bug: 68667
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474707
Original-Revision: 1cc53a6c4ea17ae859f46a531e127226c0b36be1
GitOrigin-RevId: 1cf4dd2e3ed5d40ffc59f84cc91d297a1ad93199
Change-Id: I82e37242e7d59feeb65a6e7b3b33db15cff1366f