[roll] Roll infra/recipes [test_owners] Fetch test names directly from BQ
Instead of downloading a manifest uploaded by a Cloud Function. This
will simplify the architecture of flake fetcher's owner-assignment
feature since we'll only need to maintain the test-owners builder and
can we can delete the Cloud Function entirely.

To run the query, we use the new `bigquery` tool that I wrote in
fxrev.dev/480699.

Also add a dry-run mode to the recipe to make it easier to test without
modifying the version of the test owners file that's in production.

Test: https://ci.chromium.org/swarming/task/5186589ab3cb2d10?server=chrome-swarming.appspot.com

Original-Bug: 66274
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/481217
Original-Revision: bd87d5f8fa147d3b31a3c5be4c1ee7d6cf413df8
GitOrigin-RevId: 90db63835c8020fa4c662c52c8dea7dcbacae259
Change-Id: Iec3e0a7946abe5075bd9d4dab2dd1b7c19c2b7b9