[roll] Roll fuchsia [fuchsia-hyper] Add RFC8446 Appendix C.4 session cache
TLS 1.3 makes some large changes around how sessions are resumed. In
particular, tickets are no longer sent from the server in plaintext.
While the client does send these to the server in plaintext, section
4.6.1 of the RFC suggests servers send multiple tickets to the client.

One reason for this is to prevent tracking of client sessions. Because
client session tickets are sent in plain text, it is possible to know
which sessions are related. Clients that never reuse a session ticket
cannot be tracked in this fashion.

The default session storage implementation has some properties that
preclude it from qualifying as C.4-capable:

 * It evicts tickets in an arbitrary order. When the capacity of the
   cache is full, the newest item may be the one evicted. This may later
   return the original item.
 * It only stores a single ticket, which prevents effective racing of
   TLS connections in a Happy Eyeballs implementation.
 * When a negotiation attempt fails, the ticket will be retried.

This change adds a `StoresClientSessions` implementation that addresses
these issues. The cache is sized and LRU-managed. Because keyspace is
shared between TLS 1.3 and TLS 1.2 resumption information, insertion of
one where the other is present evicts all the rest. It is expected to be
unlikely to run into cases where a SN is served by native 1.3 and 1.2
servers, and that this will continue to be unlikelier.

TLS 1.3 cache entries store 6 tickets per subject name; this value is
derived from Appendix C.4, which suggests that this is an expected
concurrency level for HTTP/1.1 clients. This is also a reasonable number
of tickets to store from a Happy Eyeballs perspective, allowing
consumption of all 6 tickets within the "magical" 2 second window
defined-in and oft-cited-from [Miller68].

This change makes the new C.4 compliant cache the default for
fuchsia_hyper clients, and updates the httpsdate client library to use
this cache as well.

[Miller68]: https://dl.acm.org/doi/10.1145/1476589.1476628

Test: New unit tests, OTA tests, CQ, bogo test in upstream.
Original-Fixed: 68871

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/478362
Original-Revision: d5f5f63a2509d089e37e4c1d670b98c3659c4a5b
GitOrigin-RevId: 9065269974ead1278492882d4312ab2f837fe833
Change-Id: I081284bf6d7ef16f9b84a2d5c4d634a985b9dd69