[roll] Roll fuchsia [bt][a2dp-source-integration-tests] Re-enable tests
Context: After the A2DP component merge, the a2dp source integration
tests started to flake. The exact flake is seen in the attempt to
connect to the cobalt.LoggerFactory service. Occasionally, tests that
were run in parallel would fail to connect to the LoggerFactory service, resulting
in an ERROR severity log and therefore failing the test. From an initial
investigation, it looks like test #1 is able to connect to the service,
but when test #2 tries shortly after, it fails due to a PEER_CLOSED on
it's ClientEnd of the FIDL channel. This could be because the ServerEnd
is getting dropped somehow. The mock_cobalt component is capable of
supporting multiple clients, so the parallelism in that regard should be
ok.

- Update manifest to use the new system services.

- As a temporary measure to re-enable the tests, limit the parallelism
of the tests to one thread. This eliminates the flakes mentioned above.
fxbug.dev/69212 tracks the work around debugging why concurrent service
accesses are failing. The net benefit of re-enabling the integration
tests outweighs the decreased performance of the tests when limiting its
parallelism.

Original-Fixed: 67823
Test: fx test bt-a2dp-source-integration-tests

Multiply: bt-a2dp-source-integration-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/480658
Original-Revision: dd1b73187891f08b8d2a93a79b25c6d13ea3dd44
GitOrigin-RevId: 598f71cba641330108916b39babcfa0f8aa58fcd
Change-Id: I14ac1189ca7d23ea326f9451f965fe88f7db59a3